Duplicate searcher improvements.
Added try/catch to GeneratePrints() to attempt to stop more fatal errors.
Added option to merge the tags of inspected duplicates.
Added option to only keep tags that are found in custom tags library.
Fixed tag highlighting in infoboxes (highlighting if tag not on other file).
Added highlighting in italic if the tag is in the local custom tag library.
Added highlighting in italic & bold simultaneously if both conditions are true.