Shortcut modifications.
Shortcuts dictionary actions are now Funcs so they can return a bool of whether the shortcut was successful or not.
If a shortcut was not successful a windows beep is played.
Multiple methods used by shortcuts have been reworked to return bool now to see if they were successful.
Left/Right arrow keys to skip through files now works in notesTextBox, tagSearchTextBox and subfolderSearchTextBox.
Inadvertently added a safety check to Undo() and Redo() that check if the stack is not empty before continuting