#4 DupeSearch - Remove reverse-dupe entry check and self-duplicate checks, optimise loading of thumbnails, made duplicates additions threadsafe
Removed for when a duplicate was found that the duplicate didnt already exist but in reverse (1&2 vs 2&1).
Removed check when finding a duplicate that it wasnt checking against itself (this is fixed by adding the print to the prints list *after* checking it for dupes against the list).
Adding to the list of duplicates is now threadsafe, stopping many collection-altered exceptions from occuring.
FilePrint now has Size computed property which gets the size when accessed.
Streamlined loading thumbnail code.
Removed round-about code where when creating a thumbnail for an image it would save the thumbnail and then load it back in from disk,
Added `SetVideoSize()` for DupeSearchForm to set video size from VLC.
Changed ffmpeg to take thumbnail from 2 seconds into the video to reduce completely black thumbnails.